add minimal problems found in alpha
