+++ packages/playground/package.json
@@ -16,8 +16,7 @@
     "react-dom": "^16.13.1",
     "react-error-boundary": "^2.3.1",
     "react-scripts": "3.4.1",
-    "rxjs": "^7.0.1",
-    "rxjs-traces": "^0.2.0-alpha.1"
+    "rxjs": "^7.0.1"
   },
   "scripts": {
     "start": "react-scripts start",
