+++ packages/playground/src/App.tsx
@@ -1,64 +1,41 @@
-import { bind, Subscribe } from "@react-rxjs/core";
 import React from "react";
-import { ErrorBoundary, FallbackProps } from "react-error-boundary";
-import { interval, merge, timer } from "rxjs";
+import { EMPTY, timer } from "rxjs";
 import { addDebugTag } from "rxjs-traces";
 import { DevTools } from "rxjs-traces-devtools";
 import "rxjs-traces-devtools/dist/bundle.css";
-import { scan, share, switchMap, withLatestFrom } from "rxjs/operators";
+import { share, switchMap } from "rxjs/operators";
 import "./App.css";
 
-const loggedUser = timer(100).pipe(share());
-const tradingAccount = loggedUser.pipe(
-  switchMap(() => timer(100)),
-  share(),
-  addDebugTag("tradingAccount")
-);
-const userViewConfig = loggedUser.pipe(addDebugTag("userViewConfig"));
-const viewConfig = tradingAccount.pipe(
-  withLatestFrom(userViewConfig),
-  addDebugTag("viewConfig")
-);
-const selectedAccount = tradingAccount.pipe(addDebugTag("selectedAccount"));
+const loggedUser = timer(1000).pipe(addDebugTag("loggedUser"), share());
+// Problem number 1. Only happens if loggedUser is async and shared
+// This top-level subscription causes loggedUser to be disconnected from any dependant.
+loggedUser.subscribe();
 
-const seconds = interval(1000).pipe(addDebugTag("seconds"), share());
-const value = seconds.pipe(
-  scan((total, s) => total + s),
-  addDebugTag("value")
+const tradingAccount = loggedUser.pipe(addDebugTag("tradingAccount"));
+
+const viewConfig = tradingAccount.pipe(addDebugTag("viewConfig"));
+
+const positionsMap = tradingAccount.pipe(
+  // Problem number 2. Happens when loggedUser emits the value
+  // 'positionsMap' should have a dependency to '1' and 'tradingAccount' (OK)
+  // but a dependency from 'viewConfig' to '1' also appears! (ERR)
+  switchMap(() => EMPTY.pipe(addDebugTag("1"))),
+  addDebugTag("positionsMap")
 );
 
-const [useStream] = bind(value);
-
-const RandomComponent = () => {
-  const value = useStream();
-  return <div>{value}</div>;
-};
-
-function ErrorFallback({
-  error,
-  componentStack,
-  resetErrorBoundary,
-}: FallbackProps) {
-  return (
-    <div role="alert">
-      <p>Something went wrong:</p>
-      <pre>{error?.message}</pre>
-      <pre>{componentStack}</pre>
-      <button onClick={resetErrorBoundary}>Try again</button>
-    </div>
-  );
-}
-
-const dep = merge(viewConfig, selectedAccount);
+// Problem 3. Maybe related/same to [1]
+// Removing this setTimeout causes everything to be disconnected
+setTimeout(() => {
+  viewConfig.subscribe();
+  positionsMap.subscribe();
+});
 
 function App() {
   return (
     <div className="App">
-      <ErrorBoundary FallbackComponent={ErrorFallback}>
-        <Subscribe fallback={null} source$={dep}>
-          <RandomComponent />
-        </Subscribe>
-      </ErrorBoundary>
+      {/* <Subscribe fallback={null} source$={dep}> */}
+      Hey
+      {/* </Subscribe> */}
       <DevTools />
     </div>
   );
